requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
psycopg2-binary
asyncpg